requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
Pillow>=8.0.0
tqdm>=4.62.0
tensorboardX>=2.4
timm>=0.5.0
ftfy>=6.0.0
regex>=2021.8.0
openpyxl>=3.0.0
